getting classTypes from server
